Animal Decimal

+++ models/Animal.js
@@ -231,7 +231,7 @@ Animal.init(
       comment: "รหัสสายพันธุ์ที่ 1",
     },
     AnimalBreedPercent1: {
-      type: DataTypes.FLOAT(),
+      type: DataTypes.DECIMAL(10,4),
       allowNull: true,
       comment: "สัดส่วนสายพันธุ์ที่ 1",
     },
@@ -241,7 +241,7 @@ Animal.init(
       comment: "รหัสสายพันธุ์ที่ 2",
     },
     AnimalBreedPercent2: {
-      type: DataTypes.FLOAT(),
+      type: DataTypes.DECIMAL(10,4),
       allowNull: true,
       comment: "สัดส่วนสายพันธุ์ที่ 2",
     },
@@ -251,7 +251,7 @@ Animal.init(
       comment: "รหัสสายพันธุ์ที่ 3",
     },
     AnimalBreedPercent3: {
-      type: DataTypes.FLOAT(),
+      type: DataTypes.DECIMAL(10,4),
       allowNull: true,
       comment: "สัดส่วนสายพันธุ์ที่ 3",
     },
@@ -261,7 +261,7 @@ Animal.init(
       comment: "รหัสสายพันธุ์ที่ 4",
     },
     AnimalBreedPercent4: {
-      type: DataTypes.FLOAT(),
+      type: DataTypes.DECIMAL(10,4),
       allowNull: true,
       comment: "สัดส่วนสายพันธุ์ที่ 4",
     },
@@ -271,7 +271,7 @@ Animal.init(
       comment: "รหัสสายพันธุ์ที่ 5",
     },
     AnimalBreedPercent5: {
-      type: DataTypes.FLOAT(),
+      type: DataTypes.DECIMAL(10,4),
       allowNull: true,
       comment: "สัดส่วนสายพันธุ์ที่ 5",
     },
@@ -281,7 +281,7 @@ Animal.init(
       comment: "url ภาพประจำตัวสัตว์",
     },
     AnimalBornWeight: {
-      type: DataTypes.FLOAT(),
+      type: DataTypes.DECIMAL(10,4),
       allowNull: true,
       comment: "น้ำหนักแรกเกิด",
     },
